Rename routes folder to stick with READ ME instructions
